Fix #353, correct PSP module IDs and lookups
Correct the CFE_PSP_Module_FindByName and CFE_PSP_Module_GetAPIEntry
lookups to use the correct length of the config data 'PspModuleList'.

The internal variable that had been used reflected the total across
two lists and thus was not a correct limit.

Rather than having these functions search both lists, this instead
keeps it to only exposing the user-specified modules in PspModuleList,
it does not expose the built-in/base modules.  This makes the change
simpler, as typically the built in modules would not be accessed
through these functions at all.